FPS_MachineLearning 1.5
FPS_MachineLearning 1.5 (Kevin Shimizu) - Refactored code to improve clarity and efficiency and remove redundancies. The algorithm still does not optimize properly for the initial tests. Need to analyze all the functions associated with training. It is likely that something is wrong with the activation functions, cost function or minimizing function.