FPS_MachineLearning 5.1
FPS_MachineLearning 5.1 (Kevin Shimizu) - Added ability to change activation functions from admin menu and adding a new layer will automatically set an activation function. Also, added the ability to change the folder networks and settings are saved to from the admin menu. Added the distance from the objective (both x and z) as an input and changed the reward structure to promote the agent moving to an objective with a random location.