FPS_MachineLearning 5.5
Added a turret to shoot projectiles at the agent. When hit by the projectiles it will deduct from the reward. The projectile's x and z distance from the agent is used as inputs into the neural network (The same way the objective distances are used). Beginning training. Still need to fully reset (destroy projectile, reset turret timer) when the episode changes.