FPS_MachineLearning 5.51
Projectile fully resets after each episode (turret timer resets and projectile is destroyed).  Agent is now rewarded for not being hit by projectile (either projectile timer elapses or projectile hits another object and is destroyed). Currently setup for a full test of "find the objective" and "avoid projectiles." Next update I am going to add a boolean as a neural network input (cast as an int) to indicate whether there is an active projectile or not.