FPS_MachineLearning 6.0
Agent has been successfully trained to navigate to an objective and avoid projectiles by dodging or taking cover. However, some adjustments need to be made to finetune behaviour i.e. increase objective reward to match projectile hit penalty so the agent must decide which is better. Added raycasts to indicate if the agent is in sight of the turret (Make sure projectile does not trigger raycasts). The agent is rewarded when not in sight (behind cover). Agent is penalized when hit by the projectile (IMPORTANT NOTE: Make sure the various rewards are not too far apart i.e. rewarding an action with +1 and another with +1000 as this will likely cause poor training results.) Increased the inputs to 15: Indices 0 and 1 = X and Z axes distance from the agent to projectile. Indices 2-9 = raycast distances to obstacles (used for collision detection). Index 10 = boolean isTargetVisible. Indices 11 and 12: Agent X and Z axes positions. Indices 13 and 14 = X and Z axes distance to objective.